Clean up; fixing more compiler warnings

+++ src/mont.c
@@ -338,15 +338,15 @@ STATIC void mont_mult_p256(uint64_t *out, const uint64_t *a, const uint64_t *b, 
 
         /* p[0] = 2³²-1 */
         prod = k2 + t32[i+0];
-        t32[i+0] = prod;
+        t32[i+0] = (uint32_t)prod;
         carry = prod >> 32;
         /* p[1] = 2³²-1 */
         prod = k2 + t32[i+1] + carry;
-        t32[i+1] = prod;
+        t32[i+1] = (uint32_t)prod;
         carry = prod >> 32;
         /* p[2] = 2³²-1 */
         prod = k2 + t32[i+2] + carry;
-        t32[i+2] = prod;
+        t32[i+2] = (uint32_t)prod;
         carry = prod >> 32;
         /* p[3] = 0 */
         t32[i+3] += carry;
@@ -364,7 +364,7 @@ STATIC void mont_mult_p256(uint64_t *out, const uint64_t *a, const uint64_t *b, 
         carry |= t32[i+6] < k;
         /* p[7] = 2³²-1 */
         prod = k2 + t32[i+7] + carry;
-        t32[i+7] = prod;
+        t32[i+7] = (uint32_t)prod;
         carry = prod >> 32;
 
         for (j=8; carry; j++) {
